Handle 'current player' without juggling turn numbers
Instead of incrementing a turn number integer, then figuring out
which player maps to that number, just send the next `currentPeerId`
as a field in the state when ending the current turn. This prevents
the current user from jumping around when players join or leave.

Unfortunately if a player leaves while it's their turn, we don't
know (yet) whose turn is next, so just default to the first player.
We'll fix this though.